requirements.txt
langchain_community
langchain_core
langchain_text_splitters
langchain_openai
langchain_cohere
langchain_chroma
huggingface_hub
gradio
python-dotenv
pypdf
llama-parse 
nest_asyncio